data_logger: backends: exfat: update cluster size
For SD cards manually assign the largest recommended cluster size,
128kB. For other backends use the default selected by the file system
library.

The cluster size is not strictly related to the block size, but instead
to the size of files created on the file system. Cluster sizes that are
too small result in huge file allocation tables.

For our purposes, we know we are only creating large binary logging
file (except README.txt), so a large cluster size is preferred.

Signed-off-by: Jordan Yates <jordan@embeint.com>