epacket: interfaces: udp: require periodic data
Require the server to send data periodically to the end device. If not
received, the end device will close the socket and requery the servers
IP address via DNS. How often a packet is required and the number of
attempts before giving up are both tunable parameters.

This protects against the UDP server changing IP address for any reason
and deployed devices never detecting it.

Signed-off-by: Jordan Yates <jordan@embeint.com>