task_runner: tasks: gnss_ubx: minimum required accuracy
Add a minimum required accuracy before locations from the modem are
provided to the system. This is done to prevent locations that are
hundreds or thousands of kilometers off from being used for decisions.

Note that the information is still broadcast to inform other libraries
that the fix is running.

Signed-off-by: Jordan Yates <jordan@embeint.com>