nrf_modem_lib: optional caching of signal quality
Allow users to enable an optional fallback to the last valid queried
signal parameters. This is useful for devices that spend most of their
time in PSM mode, where parameters cannot be queried, but users still
want some indication of the signal quality. The best we can do is the
signal quality the last time we were on the network.

In addition to caching the values when a user requests them, the cache
is also automatically updated when the cell tower changes and when the
connection transitions to active, in case no user request occurs while
the connection is active.

Signed-off-by: Jordan Yates <jordan@embeint.com>