kconfig: defaults: nrf: increase LTE RX memory size
Increase the default memory allocated to the RX portion of the modem
interface. While 40kB is allocated to the modem, only 17768 bytes are
actually used by the modem in the default configuration, with an
additional 16kB used when tracing is enabled. This leaves 6808 bytes
leftover, so allocate an extra 4096 to the RX portion.

Signed-off-by: Jordan Yates <jordan@embeint.com>