epacket: interfaces: serial_rtt: delay after send
Add a small delay after sending a packet to give debugger a chance to
rea it out. This is required to support sending bursts of packets
without silently losing data due to the behaviour of `SEGGER_RTT_Write`,
which is dropping characters if the buffer is full.

A 5ms delay still enables sending ~200 packets/sec, which is still
faster than most other interfaces.

Signed-off-by: Jordan Yates <jordan@embeint.com>