infuse: states: state snapshotting
For a system that ticks the state timeouts in the task runner context
(expected to be all systems), ensure that a state with a timeout of N
is set for N iterations of the task runner. This prevents a state being
set after task iteration but before `infuse_states_tick`, and therefore
never being seen by the runner.

Signed-off-by: Jordan Yates <jordan@embeint.com>