rpc: commands: security_state: populate hardware ID
Instead of supplying 0x00 as the proof of identity, populate the
hardware ID of the device. This is just as secure as some arbitrary
random number that we generate on the server and flash to the device
(not very). It does mean that physical JTAG access is required to each
device in order to successfully impersonate the device to the server
(or a breach of server device information).

Signed-off-by: Jordan Yates <jordan@embeint.com>