lib: security: change COAP DTLS identity
Change the COAP DTLS identity from the Infuse ID as a little endian
byte array to a hex formatted string of the Infuse ID. This is required
as the Nordic LTE modem library reads the identity as a string, and
therefore terminates the identity on the first 0x00 character. This
causes the incorrect identity to be presented to the server if any of
the ID bytes are 0x00.

Signed-off-by: Jordan Yates <jordan@embeint.com>