subsys: validation: button: added
Added validation tests for button (and other similar gpio interrupt)
driven components.
Validation can test for a rising, falling or both edges based on flags.
Because this is something that likely requires manual intervention, a
timeout in seconds for the test to last can be specified using
`INFUSE_VALIDATION_BUTTON_ACTION_TIMEOUT`.
As such, this test is off by default and can be activated using the
`INFUSE_VALIDATION_BUTTON_REQUIRE_MANUAL` config.

Signed-off-by: Aeyohan Furtado <aeyohan@embeint.com>