gnss: u_blox: swap `tx-ready` to active low polarity
Swap `tx-ready` from active high to active low polarity. This is done
as all UBlox PIO pins have a default pull-up resistor which controls
the IO state in backup modes. By having the inactive state of the pins
as high, in theory we prevent transitions on the pins when transitioning
between active and backup modes.

Signed-off-by: Jordan Yates <jordan@embeint.com>