subsys: task_runner: tasks: task_battery: fix log
Fix data logging error.
`task_battery_manual_run` accepted a `tdf` argument (pointer) which was
likely intended to update the value with the sampled results.
This `tdf` arg was unused in the run function causing any tdfs logged by
battery task to be empty.
Removed the internal variable that was being used and replaced the
reference with pointer being passed in by the function argument

Signed-off-by: Aeyohan Furtado <aeyohan@embeint.com>