drivers: watchdog: increase default timeout
Increase the default timeouts of the watchdog. Several actions can
take an on the order of 10 seconds to complete (flash erase, Bluetooth
connection), and needing to periodically interrupt these to feed the
watchdog just makes the software more complex. Increasing the default
value to 30 seconds should eliminate the vast majority of these
problems. 30 seconds is still basically instantaneous from the
perspective of the device lifetime.

Signed-off-by: Jordan Yates <jordan@embeint.com>