version: move implementation to source file
Move the `application_version_get` implementation to a source file. This
enables us to declare the version structure in ROM, which is important
for minimising the differences between binaries when the version number
changes.

Previously the number of instructions generated would depend on the
version number, with consequences for the flash layout in the remainder
of the application.

Signed-off-by: Jordan Yates <jordan@embeint.com>