task_runner: extract `struct k_thread` from data
`struct k_thread` is a large object (>= 200 bytes), which for the
majority of tasks (workqueue items) is not needed. Pull the structure
out to a pointer on the config struct to avoid the RAM overhead on
tasks that do not use it.

An added bonus is that the task structure is now its own variable in the
`.elf` file, which can help crash analysis tools (Memfault).

Signed-off-by: Jordan Yates <jordan@embeint.com>