rpc: commands: mem_read: added
Add an RPC for reading arbitrary memory addresses on a device. This is
intended as a last-resort debugging mechanism, and as such no address
validation is performed on the requested addresses. This enables reading
memory such as RAM, ROM, and peripheral registers, but means that
invalid accesses WILL fault the device.

Signed-off-by: Jordan Yates <jordan@embeint.com>