security: explicitly set cached device ID for test creds
The cached device ID the cloud side expects for `test_shared_secret`
is dependent on the random key pair generated by the tests. This in
turn can change as the entropy drivers are updated, causing the cloud
to no longer identify `test_shared_secret` properly.

Override the cached device ID to the value the cloud expects to isolate
the tests from the underlying entropy driver.

Signed-off-by: Jordan Yates <jordan@embeint.com>