epacket: interfaces: common_crypt: tx decrypt
Add a function to decrypt packets that we encrypted for transmission to
enable TX failure handlers to revert the payloads to their original
state. Assuming that transmission failures are rare, this is more
efficient than keeping an unencrypted version of the packet around until
we can free the encrypted buffer.

Signed-off-by: Jordan Yates <jordan@embeint.com>