epacket: handlers: `RPC_RSP` trigger immediate flush
Update `receive_forward` to automatically trigger a flush of the pending
buffer if the appended buffer is `RPC_RSP`. For RPC responses, it is
more important to get the response to the backend (cloud/serial) as fast
as possible so that there is a better chance the connection can be
reused. There is also the expectation that since the command has
finished, there shouldn't be as much data on the link so we are less
concerned with efficiency.

Signed-off-by: Jordan Yates <jordan@embeint.com>