epacket: packet: rate limit request packet
Define a special rate limit request packet, which is not encrypted and
simply requests the receiving device to hold off on sending any more
packets for a short duration. This is intended for use by gateways that
are acting as a pipe for bulk data between the cloud and some other
device, and therefore cannot use the more fully featured
`INFUSE_RPC_DATA_ACK` mechanism, since it is not in control of the RPC
flow.

Signed-off-by: Jordan Yates <jordan@embeint.com>