rpc: commands: data_logger_read_available: added
Add a version of `data_logger_read` with improved semantics for the
cloud data sync. Instead of specifying the first and last blocks the
earliest desired block and block count are specified. This allows the
cloud to simply specify the last block that it has downloaded and the
device will respond with the oldest data still remaining on the device,
without needing to run a `data_logger_state` command first.

Signed-off-by: Jordan Yates <jordan@embeint.com>