epacket: interfaces: udp: configurable buffer claim timeout
Make the buffer claim timeout configurable and extend the default value.
This is intended to be a worst case scenario when all other watchdogs
have failed to operate. Asserting from watchdogs etc is preferable
because they provide more information about the root cause of the
blockage, instead of just pointing at a symptom.

Signed-off-by: Jordan Yates <jordan@embeint.com>