rpc: commands: common_file_actions: split union fields
Remove the `union` parameter for the flash area and client context
fields. While they are never used together, there is logic that always
checks the value of `fa`, which is incorrect for the BT controller
related actions. Reworking the logic would result in a more complicated
solution, and the cost of the split is only 4 bytes of stack RAM.

Signed-off-by: Jordan Yates <jordan@embeint.com>