epacket: interface: udp: LTE RAI option
Add the ability to set `NRF_RAI_LAST` Release Assistance Indication
socket option for UDP. This option signals to the LTE tower that no
more transmissions are expected from the device.

If this is not true, there may be a time/data/power overhead due to
transitioning back to RRC active mode.

Signed-off-by: Jordan Yates <jordan@embeint.com>